service/src/requirements.txt
pwntools
ipython
pycryptodomex
click
